+++ ios/MullvadSettings/RecentConnectionsRepository.swift
@@ -5,6 +5,8 @@
 //  Created by Mojgan on 2025-10-15.
 //  Copyright © 2025 Mullvad VPN AB. All rights reserved.
 //
+
+import Combine
 import MullvadTypes
 
 public enum RecentConnectionsRepositoryError: LocalizedError, Hashable {
@@ -18,55 +20,105 @@ public enum RecentConnectionsRepositoryError: LocalizedError, Hashable {
     }
 }
 
-final class RecentConnectionsRepository: RecentConnectionsRepositoryProtocol {
+final public class RecentConnectionsRepository: RecentConnectionsRepositoryProtocol {
     private let store: SettingsStore
     private let maxLimit: UInt
+    private let recentConnectionsSubject: PassthroughSubject<RecentConnectionsResult, Never> = .init()
 
     private let settingsParser: SettingsParser = {
         SettingsParser(decoder: JSONDecoder(), encoder: JSONEncoder())
     }()
 
-    init(store: SettingsStore, maxLimit: UInt = 50) {
+    public var recentConnectionsPublisher: AnyPublisher<RecentConnectionsResult, Never> {
+        recentConnectionsSubject.eraseToAnyPublisher()
+    }
+
+    public init(store: SettingsStore, maxLimit: UInt = 50) {
         self.store = store
         self.maxLimit = maxLimit
     }
 
-    func setRecentsEnabled(_ isEnabled: Bool) throws {
-        // Clear all recents whenever the recents feature status changes.
-        try write(RecentConnections(isEnabled: isEnabled, entryLocations: [], exitLocations: []))
+    public func disable() {
+        do {
+            // Clear all recents whenever the recents feature status changes.
+            let value = RecentConnections(isEnabled: false, entryLocations: [], exitLocations: [])
+            try write(value)
+            recentConnectionsSubject.send(.success(value))
+        } catch {
+            recentConnectionsSubject.send(.failure(error))
+        }
+    }
+    public func enable(_ selectedEntryRelays: UserSelectedRelays?, selectedExitRelays: UserSelectedRelays) {
+        do {
+            // Enable recents with the last selected locations for entry and exit.
+            let value = RecentConnections(
+                entryLocations: (selectedEntryRelays != nil) ? [selectedEntryRelays!] : [],
+                exitLocations: [selectedExitRelays])
+            try write(value)
+            recentConnectionsSubject.send(.success(value))
+        } catch {
+            recentConnectionsSubject.send(.failure(error))
+        }
     }
 
-    func add(_ location: UserSelectedRelays, as type: RecentLocationType) throws {
-        let current = try read()
-        guard current.isEnabled else { throw RecentConnectionsRepositoryError.recentsDisabled }
-        var currentList = current[keyPath: keyPath(for: type)]
-        if let idx = currentList.firstIndex(of: location) { currentList.remove(at: idx) }
-        currentList.insert(location, at: 0)
-        currentList = Array(currentList.prefix(Int(maxLimit)))
+    public func add(_ selectedEntryRelays: UserSelectedRelays, selectedExitRelays: UserSelectedRelays) {
+        do {
+            let current = try read()
+            guard current.isEnabled else { throw RecentConnectionsRepositoryError.recentsDisabled }
 
-        let new =
-            (type == .entry)
-            ? RecentConnections(
-                isEnabled: current.isEnabled, entryLocations: currentList, exitLocations: current.exitLocations)
-            : RecentConnections(
-                isEnabled: current.isEnabled, entryLocations: current.entryLocations, exitLocations: currentList)
+            let insertAtZero: ([UserSelectedRelays], UserSelectedRelays) -> [UserSelectedRelays] = { recents, recent in
+                var currentRecents = recents
 
-        try write(new)
+                currentRecents.removeAll(where: { item in
+                    let recentCustomList = recent.customListSelection
+                    let itemCustomList = item.customListSelection
+
+                    // Both have custom lists & same listId
+                    if let recentCL = recentCustomList,
+                        let itemCL = itemCustomList,
+                        recentCL.listId == itemCL.listId
+                    {
+                        // If both represent the list itself, then replace the old item with the new one.
+                        if recentCL.isList && itemCL.isList {
+                            return true
+                        }
+                        // Otherwise remove only if the two entries match exactly
+                        return recent == item
+                    }
+
+                    // Locations match and neither entry is a list node
+                    if recent.locations == item.locations {
+                        return !(recentCustomList?.isList == true || itemCustomList?.isList == true)
+                    }
+
+                    // No match
+                    return false
+                })
+                currentRecents.insert(recent, at: 0)
+                return Array(currentRecents.prefix(Int(self.maxLimit)))
+            }
+            let new = RecentConnections(
+                entryLocations: insertAtZero(current.entryLocations, selectedEntryRelays),
+                exitLocations: insertAtZero(current.exitLocations, selectedExitRelays))
+            try write(new)
+            recentConnectionsSubject.send(.success(new))
+
+        } catch {
+            recentConnectionsSubject.send(.failure(error))
+        }
     }
 
-    func all() throws -> RecentConnections {
-        try read()
+    public func initiate() {
+        do {
+            let value = try read()
+            recentConnectionsSubject.send(.success(value))
+        } catch {
+            recentConnectionsSubject.send(.failure(error))
+        }
     }
 }
 
 private extension RecentConnectionsRepository {
-    private func keyPath(for type: RecentLocationType) -> KeyPath<RecentConnections, [UserSelectedRelays]> {
-        switch type {
-        case .entry: return \.entryLocations
-        case .exit: return \.exitLocations
-        }
-    }
-
     private func read() throws -> RecentConnections {
         let data = try store.read(key: .recentConnections)
         return try settingsParser.parseUnversionedPayload(as: RecentConnections.self, from: data)