+++ ios/MullvadVPN/Classes/AccessbilityIdentifier.swift
@@ -73,6 +73,8 @@ public enum AccessibilityIdentifier: Equatable {
     case locationListItem(String)
     case entryLocationButton
     case exitLocationButton
+    case recentConnectionsToggleButton
+    case disableRecentConnectionsButton
 
     // Cells
     case deviceCell