+++ ios/MullvadVPNTests/MullvadVPN/View controllers/SelectLocation/AllLocationsDataSourceTests.swift
@@ -96,90 +96,57 @@ class AllLocationsDataSourceTests: XCTestCase {
         XCTAssertEqual(nodeByLocation, nodeByCode)
     }
 
-    func testConnectedNode() throws {
-        let hostname = "es1-wireguard"
-        dataSource.setConnectedRelay(hostname: hostname)
-        dataSource.nodes.forEachNode { node in
-            XCTAssertEqual(node.isConnected, node.name == hostname)
-        }
-
-        dataSource.setConnectedRelay(hostname: "invalid-hostname")
-        dataSource.nodes.forEachNode { node in
-            XCTAssertFalse(node.isConnected)
-        }
-    }
-
-    func testSetSelectedLocation() throws {
-        dataSource.setSelectedNode(selectedRelays: .init(locations: [.country("es")]))
-
-        dataSource.nodes.forEachNode { node in
-            if node.locations == [.country("es")] {
-                XCTAssertTrue(node.isSelected)
-            } else {
-                XCTAssertFalse(node.isSelected)
-            }
-        }
-
-        dataSource
-            .setSelectedNode(
-                selectedRelays: .init(locations: [.country("invalid")])
-            )
-        dataSource.nodes.forEachNode { node in
-            XCTAssertFalse(node.isSelected)
-        }
-    }
-
-    func testDoNotSetSelectedCustomListLocation() throws {
-        let selectedRelays: UserSelectedRelays = .init(
-            locations: [
-                .country("es")
-            ],
-            customListSelection: .init(listId: .init(), isList: false)
-        )
-
-        dataSource.setSelectedNode(selectedRelays: selectedRelays)
-
-        dataSource.nodes.forEachNode { node in
-            XCTAssertFalse(node.isSelected)
-        }
-    }
-
-    func testExcludeLocation() throws {
-        let excludedRelays = UserSelectedRelays(locations: [.hostname("se", "sto", "se2-wireguard")])
-        dataSource.setExcludedNode(excludedSelection: excludedRelays)
-        let excludedNode = dataSource.node(by: excludedRelays)!
-
-        XCTAssertTrue(excludedNode.isExcluded)
-
-        excludedNode.forEachAncestor { ancestor in
-            XCTAssertFalse(ancestor.isExcluded)
-        }
-
-        let includedNode = dataSource.node(by: .init(locations: [.country("es")]))!
-        XCTAssertFalse(includedNode.isExcluded)
-        includedNode.forEachDescendant { child in
-            XCTAssertFalse(child.isExcluded)
-        }
-    }
-
-    func testExcludeLocationIncludesAncestors() throws {
-        let excludedRelays = UserSelectedRelays(locations: [.hostname("es", "mad", "es1-wireguard")])
-        dataSource.setExcludedNode(excludedSelection: excludedRelays)
-        let excludedNode = dataSource.node(by: excludedRelays)!
-
-        XCTAssertTrue(excludedNode.isExcluded)
-
-        // All ancestors are exluded when single child is excluded
-        excludedNode.forEachAncestor { ancestor in
-            XCTAssertTrue(ancestor.isExcluded)
-        }
-
-        let includedNode = dataSource.node(by: .init(locations: [.country("se")]))!
-        XCTAssertFalse(includedNode.isExcluded)
-        includedNode.forEachDescendant { child in
-            XCTAssertFalse(child.isExcluded)
-        }
-    }
+    //    func testConnectedNode() throws {
+    //        let hostname = "es1-wireguard"
+    //        dataSource.setConnectedRelay(hostname: hostname)
+    //        dataSource.nodes.forEachNode { node in
+    //            XCTAssertEqual(node.isConnected, node.name == hostname)
+    //        }
+    //
+    //        dataSource.setConnectedRelay(hostname: "invalid-hostname")
+    //        dataSource.nodes.forEachNode { node in
+    //            XCTAssertFalse(node.isConnected)
+    //        }
+    //    }
+    //
+    //    func testExcludeLocation() throws {
+    //        let excludedRelays = UserSelectedRelays(locations: [.hostname("se", "sto", "se2-wireguard")])
+    //        dataSource.setExcludedNode(excludedSelection: excludedRelays)
+    //        let excludedNode = dataSource.node(by: excludedRelays)!
+    //
+    //        XCTAssertTrue(excludedNode.isExcluded)
+    //
+    //        excludedNode.forEachAncestor { ancestor in
+    //            XCTAssertFalse(ancestor.isExcluded)
+    //        }
+    //
+    //        let includedRelays = UserSelectedRelays(locations: [.country("es")])
+    //        let includedNode = dataSource.node(by: includedRelays)!
+    //        XCTAssertFalse(includedNode.isExcluded)
+    //        includedNode.forEachDescendant { child in
+    //            XCTAssertFalse(child.isExcluded)
+    //        }
+    //    }
+    //
+    //    func testExcludeLocationIncludesAncestors() throws {
+    //        let excludedRelays = UserSelectedRelays(locations: [.hostname("es", "mad", "es1-wireguard")])
+    //        dataSource.setExcludedNode(excludedSelection: excludedRelays)
+    //        let excludedNode = dataSource.node(by: excludedRelays)!
+    //
+    //        XCTAssertTrue(excludedNode.isExcluded)
+    //
+    //        // All ancestors are exluded when single child is excluded
+    //        excludedNode.forEachAncestor { ancestor in
+    //            XCTAssertTrue(ancestor.isExcluded)
+    //        }
+    //
+    //        let includedRelays = UserSelectedRelays(locations: [.country("se")])
+    //        let includedNode = dataSource.node(by: includedRelays)!
+    //        XCTAssertFalse(includedNode.isExcluded)
+    //        includedNode.forEachDescendant { child in
+    //            XCTAssertFalse(child.isExcluded)
+    //        }
+    //    }
 }
 
 extension AllLocationsDataSourceTests {